Fixed #26052 -- Moved conditional_content_removal() processing to the test client.
